backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-CORS==4.0.1
APScheduler==3.11.0
gunicorn==21.2.0
redis==5.3.0
minio==7.2.15
prophet==1.1.7
openai==0.28.1
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.2
plotly==6.2.0
PyMySQL==1.1.1
paho-mqtt==1.6.1
requests==2.32.4
asyncio-mqtt==0.16.2
python-dotenv==1.1.1
aiohttp==3.12.13
cryptography==41.0.7

# 告警系统依赖
bcrypt==4.0.1
email-validator==2.0.0